usize_set
Trait IndexSet

pub trait IndexSet {
    // Required methods
    fn len(&self) -> usize;
    fn is_empty(&self) -> bool;
    fn insert(&mut self, index: usize);
    fn remove(&mut self, index: usize);
    fn contains(&self, index: usize) -> bool;
    fn iter(&self) -> impl Iterator<Item = usize> + '_;
    fn union(&mut self, other: &Self);

    // Provided method
    fn reserve(&mut self, _size: usize) { ... }
}

Public interface of any index set implementation.

Required Methods

fn len(&self) -> usize

Return the number of usize values present in this IndexSet.

fn is_empty(&self) -> bool

Checks if this IndexSet has no inner indexes stored within.

fn insert(&mut self, index: usize)

Add a new index to this IndexSet.

fn remove(&mut self, index: usize)

Remove an index from this IndexSet.

fn contains(&self, index: usize) -> bool

Check the presence of an index in this IndexSet.

fn iter(&self) -> impl Iterator<Item = usize> + '_

Return an iterator over the indices in this IndexSet, in ascending order.

fn union(&mut self, other: &Self)

Merge two IndexSet instances.

Corresponds to a mutating set union operation, between self and other.

Provided Methods

fn reserve(&mut self, _size: usize)

Attempt to reserve space for the specified number of additional usize elements.

Object Safety

This trait is not object safe.